Prices and Multipliers Added, Charging player fixed, Build and Show Commands updated

+++ Ultimate City Building Simulator/Game/Building/IBuildable.cs
@@ -9,6 +9,6 @@ namespace UltimateCityBuildingSimulator.Game.Building
     public interface IBuildable : ICloneable
     {
         int ProvideCost();
-        int ProvideMultiplier();
+        float ProvideMultiplier();
     }
 }
